Fix numeric tooltips on X on horizontal charts
The problem was that we were changing the tooltip value to the tick's label.
The change was to only do that if the tick's label or value (depending if the
axis is in categories mode or not) is equal to the x value. If it's not, we
don't replace it, which will make the correct part of the code (using
tickFormatter) to be ran.

This fixes #74.